notification-service/requirements.txt
flask
datetime
psycopg2-binary
boto3
